Combining sync and async thrift clients
