+++ src/app/beer_garden/api/http/handlers/v1/admin.py
@@ -1,6 +1,6 @@
 from beer_garden.api.http.authorization import check_permission, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
-from beer_garden.api.http.thrift import ThriftClient
+from beer_garden.api.thrift.client import ThriftClient
 from brewtils.errors import ModelValidationError
 from brewtils.schema_parser import SchemaParser
