+++ src/app/beer_garden/api/http/handlers/v1/command.py
@@ -1,6 +1,6 @@
 from beer_garden.api.http.authorization import authenticated, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
-from beer_garden.api.http.thrift import ThriftClient
+from beer_garden.api.thrift.client import ThriftClient
 
 
 class CommandAPI(BaseHandler):
