+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -3,7 +3,7 @@ import json
 import beer_garden.bg_utils
 from beer_garden.api.http.authorization import authenticated, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
-from beer_garden.api.http.thrift import ThriftClient
+from beer_garden.api.thrift.client import ThriftClient
 from brewtils.errors import ModelValidationError, RequestPublishException
 from brewtils.models import Request
 from brewtils.schema_parser import SchemaParser
