+++ src/app/beer_garden/api/http/handlers/v1/system.py
@@ -1,7 +1,7 @@
 import beer_garden.bg_utils
 from beer_garden.api.http.authorization import authenticated, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
-from beer_garden.api.http.thrift import ThriftClient
+from beer_garden.api.thrift.client import ThriftClient
 from brewtils.errors import ConflictError
 from brewtils.schemas import SystemSchema
