+++ src/app/beer_garden/api/http/handlers/v2/job.py
@@ -6,7 +6,7 @@ from brewtils.schemas import JobSchema
 
 from beer_garden.api.http.authorization import authenticated, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
-from beer_garden.api.http.thrift import ThriftClient
+from beer_garden.api.thrift.client import ThriftClient
 
 
 class JobAPI(BaseHandler):
